+++ Runtime Components/AssetBundle/Render_Texture.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 04b1f456641b4e749948d1de5652ef15
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
