+++ Runtime Components/AssetBundle/Simple_Scene.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b7620a1bb69f5d2459fad8aaf7816669
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
