+++ Runtime Components/AssetBundle/Video_Player.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a97e93e2de35c7c40b9a66e70d253efd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
