--- Runtime Components/GUI_Panel.cs
@@ -1,96 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-
-public class GUI_Panel : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-       if(Canvas == null)
-       {
-          Debug.LogError("Canvas cannot be null");
-       }
-       else 
-       {
-Texture2D Backgroundtex = new Texture2D(2, 2);
-byte[] BackgroundSprite = new byte[] 
-{
-	0x89, 0x50, 0x4E, 0x47, 0x0D, 0x0A, 0x1A, 0x0A, 0x00, 0x00, 0x00, 0x0D,
-	0x49, 0x48, 0x44, 0x52, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x20,
-	0x08, 0x06, 0x00, 0x00, 0x00, 0x73, 0x7A, 0x7A, 0xF4, 0x00, 0x00, 0x00,
-	0x09, 0x70, 0x48, 0x59, 0x73, 0x00, 0x00, 0x0E, 0xC4, 0x00, 0x00, 0x0E,
-	0xC4, 0x01, 0x95, 0x2B, 0x0E, 0x1B, 0x00, 0x00, 0x02, 0x77, 0x49, 0x44,
-	0x41, 0x54, 0x78, 0x9C, 0xED, 0x57, 0xBF, 0x8B, 0x1A, 0x51, 0x10, 0x1E,
-	0xDD, 0x45, 0x3D, 0x9B, 0x60, 0x21, 0x16, 0x29, 0x44, 0x38, 0x88, 0x08,
-	0x91, 0xE0, 0xFD, 0x01, 0x81, 0xA8, 0xA8, 0x55, 0x4A, 0x11, 0x54, 0x84,
-	0x83, 0xC0, 0x81, 0x10, 0x38, 0x21, 0x55, 0xCA, 0x40, 0x20, 0x90, 0x54,
-	0x81, 0x80, 0x70, 0x20, 0xA2, 0x58, 0xA7, 0x08, 0x2A, 0x4A, 0x20, 0xFF,
-	0x41, 0xF0, 0x44, 0x4C, 0x69, 0x3A, 0x0B, 0xBB, 0x60, 0x54, 0xFC, 0x91,
-	0x6F, 0x36, 0xFB, 0x96, 0x77, 0x62, 0x40, 0x3D, 0x97, 0x34, 0xFB, 0xC1,
-	0x30, 0xBA, 0xB3, 0xFB, 0xE6, 0x7B, 0xF3, 0xDE, 0xDB, 0x9D, 0x4F, 0x0D,
-	0x85, 0x42, 0x24, 0xC3, 0xE5, 0x72, 0x91, 0xDD, 0x6E, 0xD7, 0x7E, 0xAF,
-	0xD7, 0x6B, 0xCD, 0x6F, 0x36, 0x1B, 0xB2, 0xD9, 0x6C, 0x74, 0x08, 0xE4,
-	0x67, 0xE4, 0xF1, 0x66, 0xB3, 0xD9, 0x9D, 0xFB, 0x54, 0xA7, 0xD3, 0x49,
-	0xF3, 0xF9, 0x9C, 0x84, 0xE7, 0x07, 0x57, 0xAB, 0x15, 0x29, 0x8A, 0x42,
-	0xA7, 0x04, 0x8F, 0x29, 0xC0, 0x93, 0x64, 0x22, 0xEC, 0x55, 0x4E, 0xE4,
-	0x76, 0xBB, 0x69, 0x3A, 0x9D, 0x6A, 0x41, 0x29, 0xF1, 0x23, 0xD8, 0x0B,
-	0xD8, 0x53, 0xD8, 0x13, 0x0E, 0x1D, 0x9A, 0x13, 0xF6, 0x1D, 0xF6, 0x0D,
-	0x56, 0xC6, 0xB8, 0x3F, 0x44, 0x80, 0x93, 0x73, 0x4E, 0x86, 0xCA, 0xCC,
-	0xE4, 0x0A, 0xE8, 0x4C, 0xAF, 0x61, 0x6F, 0xB2, 0xD9, 0xEC, 0x59, 0x30,
-	0x18, 0x24, 0xBF, 0xDF, 0x6F, 0x94, 0x71, 0x5F, 0xA0, 0xDC, 0xCA, 0x68,
-	0x34, 0xBA, 0x18, 0x0E, 0x87, 0x17, 0xB5, 0x5A, 0xED, 0x0A, 0x97, 0x5E,
-	0xC3, 0x3E, 0x70, 0x8C, 0x73, 0xF1, 0x84, 0xD9, 0xAB, 0x9C, 0x94, 0xA1,
-	0x7B, 0x07, 0xD6, 0xED, 0x73, 0x2C, 0x16, 0x4B, 0xA6, 0x52, 0x29, 0xF2,
-	0xF9, 0x7C, 0x74, 0x2C, 0x98, 0x70, 0x20, 0x10, 0xD0, 0x2C, 0x12, 0x89,
-	0x9C, 0x35, 0x9B, 0xCD, 0xF7, 0xDD, 0x6E, 0x37, 0x8E, 0xD0, 0x73, 0xE4,
-	0x5A, 0x88, 0x9C, 0xDA, 0x1E, 0x90, 0xF0, 0x2E, 0x91, 0x48, 0x24, 0x73,
-	0xB9, 0x1C, 0x9D, 0x12, 0x3C, 0x91, 0x42, 0xA1, 0xC0, 0xCB, 0x9B, 0x6C,
-	0xB7, 0xDB, 0x6F, 0x71, 0xA9, 0x24, 0x62, 0xAA, 0x54, 0xDA, 0xC7, 0x28,
-	0x5B, 0x31, 0x93, 0xC9, 0x90, 0x59, 0x48, 0xA7, 0xD3, 0xD4, 0x6A, 0xB5,
-	0x5E, 0x62, 0xA3, 0xDF, 0xE0, 0xEF, 0x40, 0x23, 0x20, 0x8E, 0x1A, 0x70,
-	0x99, 0xCF, 0xE7, 0x15, 0x55, 0x55, 0x4D, 0x23, 0xE0, 0x70, 0x38, 0x08,
-	0xFB, 0x4A, 0xA9, 0xD7, 0xEB, 0x97, 0xA4, 0x57, 0x41, 0xCE, 0x96, 0x0C,
-	0x87, 0xC3, 0x64, 0x36, 0x38, 0x07, 0x08, 0x24, 0x0D, 0x02, 0xD2, 0x0B,
-	0x26, 0xE0, 0xF5, 0x7A, 0x4D, 0x27, 0xE0, 0xF1, 0x78, 0xF8, 0x5D, 0xF3,
-	0x50, 0xFC, 0x97, 0x2B, 0xE0, 0x30, 0xB3, 0xFC, 0x02, 0x7C, 0xFE, 0x31,
-	0xE9, 0x07, 0xBB, 0x08, 0xFC, 0x17, 0x58, 0x04, 0x2C, 0x02, 0x16, 0x01,
-	0x8B, 0x80, 0x45, 0x40, 0x26, 0xB0, 0x58, 0x2E, 0x97, 0xA6, 0x7F, 0x0F,
-	0xF4, 0xAE, 0xF8, 0xD7, 0x2E, 0x02, 0x3F, 0x27, 0x93, 0xC9, 0xF9, 0x7D,
-	0xDA, 0xB0, 0x7D, 0x80, 0x1C, 0x5A, 0xAE, 0x5D, 0x04, 0x5A, 0xBD, 0x5E,
-	0xAF, 0x18, 0x8F, 0xC7, 0x4D, 0x25, 0xD0, 0xEF, 0xF7, 0xF9, 0x73, 0xDC,
-	0x31, 0x08, 0x48, 0xFD, 0x40, 0xB9, 0x52, 0xA9, 0x5C, 0x45, 0xA3, 0x51,
-	0xE5, 0xD0, 0x0E, 0x78, 0x5F, 0x60, 0x89, 0xA9, 0x5A, 0xAD, 0xAE, 0x30,
-	0xFE, 0x8D, 0x41, 0x80, 0x85, 0x88, 0x8E, 0x5B, 0x90, 0xF9, 0x84, 0x6E,
-	0xA5, 0x78, 0xEA, 0xA6, 0x54, 0xA0, 0xD1, 0x68, 0xB0, 0xFB, 0x88, 0x36,
-	0xF0, 0xD6, 0x20, 0xB0, 0x25, 0x95, 0x4A, 0xE8, 0x5A, 0xCF, 0xA1, 0x0D,
-	0xEE, 0xDD, 0x96, 0xCB, 0x18, 0x8F, 0xC7, 0x84, 0xB6, 0x9C, 0xD0, 0x96,
-	0xB7, 0xD0, 0x8A, 0xBF, 0x92, 0x63, 0xDB, 0xD2, 0x6C, 0x81, 0x8A, 0xA4,
-	0x3A, 0x9D, 0xCE, 0x35, 0x4C, 0x13, 0x26, 0xAC, 0x1D, 0x59, 0x98, 0x1C,
-	0x03, 0x08, 0x13, 0x1A, 0x0C, 0x06, 0x04, 0x61, 0xF2, 0x9B, 0xB6, 0x84,
-	0x89, 0xC8, 0xF9, 0x2F, 0x69, 0xC6, 0x37, 0x7E, 0xC1, 0x72, 0xB0, 0x34,
-	0x7B, 0x46, 0x7F, 0xA5, 0xD9, 0x31, 0x60, 0x69, 0xF6, 0x15, 0x63, 0x96,
-	0xE1, 0x0D, 0x69, 0xC6, 0xC9, 0x0D, 0x69, 0x26, 0x14, 0xAB, 0x10, 0x8C,
-	0x42, 0x44, 0xEA, 0x5A, 0xAE, 0x44, 0x27, 0x82, 0x2C, 0x4E, 0xEF, 0x48,
-	0x33, 0xB1, 0x07, 0x84, 0xE7, 0x53, 0x61, 0xC6, 0x29, 0xE0, 0xCA, 0x0A,
-	0x0D, 0x22, 0xCB, 0xC1, 0x3F, 0xF7, 0x7C, 0x1A, 0xD3, 0x68, 0x32, 0x31,
-	0xEF, 0x00, 0x00, 0x00, 0x00, 0x49, 0x45, 0x4E, 0x44, 0xAE, 0x42, 0x60,
-	0x82
-};
-          Backgroundtex.LoadImage(BackgroundSprite);
-          GameObject Panel = new GameObject("Panel", typeof(Image));
-          Panel.transform.SetParent(Canvas.transform);
-          Panel.GetComponent<Image>().sprite = Sprite.Create(Backgroundtex, new Rect(0.0f, 0.0f, Backgroundtex.width, Backgroundtex.height), new Vector2(16f, 16f), 200.0f, 0, SpriteMeshType.Tight, new Vector4 (10,10,10,10));
-          Panel.GetComponent<Image>().type = Image.Type.Sliced;
-          Panel.GetComponent<Image>().color = colorOfTheImage;
-          Panel.GetComponent<RectTransform>().anchorMax = new Vector2(1,1);
-          Panel.GetComponent<RectTransform>().anchorMin = new Vector2(0,0);
-          Panel.GetComponent<RectTransform>().sizeDelta = new Vector2 (0,0);
-          Panel.GetComponent<RectTransform>().offsetMax = new Vector2(0,0);
-          Panel.GetComponent<RectTransform>().offsetMin = new Vector2(0,0);
-       }
-    }
-    public GameObject Canvas;
-    public Color colorOfTheImage = new Color(1f, 1f, 1f, 0.4f);
-}
