+++ Runtime Components/UI/GUI_Canvas.cs
@@ -1,0 +1,14 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class GUI_Canvas : MonoBehaviour
+{
+    // Start is called before the first frame update
+    void Awake()
+    {
+          GameObject Canvas = new GameObject("Canvas", typeof(Canvas), typeof(CanvasScaler), typeof(GraphicRaycaster));
+          Canvas.GetComponent<Canvas>().renderMode = RenderMode.ScreenSpaceOverlay;
+    }
+}
