+++ Runtime Components/UI/GUI_Canvas.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dd66cb552febd6449a31f7ea75aecd28
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
