+++ Runtime Components/UI/GUI_Scrollbar.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 196a64d3cbc4db0479b9db1bfaca0bc6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
