+++ Runtime Components/UI/GUI_slider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2c6246483fa958e429f456519a95e6aa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
